ASoC: sun8i-codec: Add support for AIF3
This adds the new DAI, clocks, widgets, and routes for AIF2.

AIF3 lacks some of the configuration of the other AIFs. Skip setting
bits that are unavailable on AIF3. It is not documented what the actual
settings for AIF3 are, so those parameters cannot be validated.

AIF3 can pull its bitclock from AIF1 or AIF2. Since the normal usage of
AIF3 is for low-bitrate mono audio, similar to AIF2, let's set the AIF3
clock to AIF2 for now.

Signed-off-by: Samuel Holland <samuel@sholland.org>