drm/panel: st7703: Improve the power up sequence of the panel
The datasheet specifies that it's better to keep reset asserted
while powering up the supplies, and that IOVCC should be enabled
first.

There also needs to be a delay after enabling the supplies and
before deasserting the reset. The datasheet specifies 1ms after
the supplies reach the required voltage. Use 10-20ms to give the
power supplies some time to reach the required voltage, too.

This fixes panel initialization failures during resume from sleep
on PinePhone.

Signed-off-by: Ondrej Jirman <megous@megous.com>