modem-power: Send KEY_WAKEUP on ring
Ubuntu Touch is unable to use the RING messages on the AT interface to
learn about incoming messages. Instead, we check for messages when the
phone's screen wakes.

Sending KEY_WAKEUP always wakes the screen on Ubuntu Touch, whereas the
IRQ events previously did not. Once the screen is awake, we receive our
messages.

This probably isn't suitable for merge to megi's upstream kernel.

# Conflicts:
#	drivers/misc/modem-power.c